console: Separate tti only parts and gateway fleet only parts
