+++ pkg/webui/locales/en.json
@@ -62,9 +62,6 @@
   "components.progress-bar.index.percentage": "{percentage, number, percent} finished",
   "components.qr-modal-button.index.scanContinue": "Please scan the QR code to continue. {qrScanDoc}",
   "components.qr-modal-button.index.apply": "Apply",
-  "components.qr-modal-button.index.fleetToken": "Fleet owner token",
-  "components.qr-modal-button.index.addFleet": "Add to Fleet",
-  "components.qr-modal-button.index.addToFleetTooltip": "You are registering a Managed gateway. If you want to add it to an existing fleet, click here.",
   "components.qr.input.capture.index.uploadImage": "Upload a Photo",
   "components.qr.input.capture.index.qrCodeNotFound": "QR code not found",
   "components.qr.input.video.index.fetchingCamera": "Waiting for camera…",
@@ -628,6 +625,9 @@
   "console.containers.gateway-onboarding-form.gateway-provisioning-form.index.continue": "To continue, please confirm the Gateway EUI so we can determine onboarding options",
   "console.containers.gateway-onboarding-form.gateway-provisioning-form.index.emtyEui": "Continue without EUI",
   "console.containers.gateway-onboarding-form.gateway-provisioning-form.index.noEui": "No gateway EUI",
+  "console.containers.gateway-onboarding-form.qr-scan-section.fleets-scan.addFleet": "Add to Fleet",
+  "console.containers.gateway-onboarding-form.qr-scan-section.fleets-scan.fleetToken": "Fleet owner token",
+  "console.containers.gateway-onboarding-form.qr-scan-section.fleets-scan.addToFleetTooltip": "You are registering a Managed gateway. If you want to add it to an existing fleet, click here.",
   "console.containers.gateway-onboarding-form.qr-scan-section.index.hasGatewayQR": "Does your gateway have a LoRaWAN® Gateway Identification QR Code? Scan it to speed up onboarding.",
   "console.containers.gateway-onboarding-form.qr-scan-section.index.gatewayGuide": "Gateway registration help",
   "console.containers.gateway-onboarding-form.qr-scan-section.index.invalidQRCode": "Invalid QR code data. Please note that only TTIGPRO1 Gateway Identification QR Code can be scanned. Some gateways have unrelated QR codes printed on them that cannot be used.",
