+++ pkg/webui/locales/ja.json
@@ -62,9 +62,6 @@
   "components.progress-bar.index.percentage": "{percentage}%　終了",
   "components.qr-modal-button.index.scanContinue": "続けるにはQRコードを読み取ってください{qrScanDoc}",
   "components.qr-modal-button.index.apply": "適用",
-  "components.qr-modal-button.index.fleetToken": "",
-  "components.qr-modal-button.index.addFleet": "",
-  "components.qr-modal-button.index.addToFleetTooltip": "",
   "components.qr.input.capture.index.uploadImage": "写真をアップロード",
   "components.qr.input.capture.index.qrCodeNotFound": "QRコードが見つかりません",
   "components.qr.input.video.index.fetchingCamera": "",
@@ -628,6 +625,9 @@
   "console.containers.gateway-onboarding-form.gateway-provisioning-form.index.continue": "",
   "console.containers.gateway-onboarding-form.gateway-provisioning-form.index.emtyEui": "",
   "console.containers.gateway-onboarding-form.gateway-provisioning-form.index.noEui": "",
+  "console.containers.gateway-onboarding-form.qr-scan-section.fleets-scan.addFleet": "",
+  "console.containers.gateway-onboarding-form.qr-scan-section.fleets-scan.fleetToken": "",
+  "console.containers.gateway-onboarding-form.qr-scan-section.fleets-scan.addToFleetTooltip": "",
   "console.containers.gateway-onboarding-form.qr-scan-section.index.hasGatewayQR": "",
   "console.containers.gateway-onboarding-form.qr-scan-section.index.gatewayGuide": "",
   "console.containers.gateway-onboarding-form.qr-scan-section.index.invalidQRCode": "",
